Listen on both ipv4 and ipv6 in nginx
For environments that are dual-stack, the current config will only listen on ipv4. Adding an extra listen directive should allow connections to be served on both ipv4 & ipv6.

I couldn't figure out how to do this with a single `listen` directive. In theory, a single `[::]:5000 ipv6only=off` should work, but trying that on my local enviroment, ipv4 clients were unable to connect. A mystery.