Clarification of installation steps with Coral accelerator
M.2/PCI-e Coral accelerators use different device names to the USB accelerator. Additionally, while Coral is mentioned in other pages, it is not explicitly clear that this is not quite 'plug and play', which will trip up novice users.

The config file argument for docker needs to be a directory, not a file.

Added some other points to help clarify based on my first experience.