Bugfix: get_min_region_size() returned half-size
All users expect the returned value to be the full model size, not the
half size. The effect of this bug was to increase the number of regions
that must be detected, reduce the context provided to the detector for
each, and force most regions to have to be scaled up, which became a CPU
bottleneck.