Add PoC implementation of batch delete
The batch delete goes through all the already displayed Events pages and
deletes any Event that is not marked as favorite/saved. It doesn't query
the database to get the full list of Events, as the user should be able
to clearly understand which Events will be cleaned up.

After the list of Events to be deleted is identified, it will prompt the
user to confirm the deletion of the number of Events. It will then show
a spinner until the Events are deleted and then show a feedback dialog
mentioning the number of Events deleted.

The button for the Cleanup operation is put on the same row as the
filters as adding a new row for a single button seemed ugly. It's marked
to be different with a border, since the Cleanup operation is not a
filtering option. The SVG was downloaded from [1] and it's licensed
as Public Domain [2].

1 - https://www.svgrepo.com/svg/494112/clean-up
2 - https://www.svgrepo.com/page/licensing/#CC0